Updated creation of header.
Updated the ccreation of the header section and made adjustments to
acccomodate for missing headings of other pages.
