+++ src/models/IHeaderObject.ts
@@ -1,0 +1,10 @@
+export interface IHeaderObject {
+    sectionTitle: {
+        h1: HTMLHeadingElement['innerHTML'],
+        p?: HTMLParagraphElement['innerHTML']
+    },
+    sectionDiscription: {
+        p1: HTMLParagraphElement['innerHTML'],
+        p2?: HTMLParagraphElement['innerHTML']
+    }
+}
